Update SessionKeysAndWsapiSystemAccess.psm1
When using this in IIS PowerShell isn't able to compile C# code on the fly. In my opinion it'd be better to let the user handle self-signed certificate issues. Possible solution would be to write an assembly for TrustAllCertsPolicy :
Add-Type -TypeDefinition @"
    using System.Net;
    using System.Security.Cryptography.X509Certificates;
    public class TrustAllCertsPolicy : ICertificatePolicy {
        public bool CheckValidationResult(
            ServicePoint srvPoint, X509Certificate certificate,
            WebRequest request, int certificateProblem) {
            return true;
        }
    }
"@ -OutputAssembly C:\Temp\TrustAllCertsPolicy.dll

and later on load it:
Add-Type -Path C:\Temp\TrustAllCertsPolicy.dll
[System.Net.ServicePointManager]::CertificatePolicy = New-Object -TypeName TrustAllCertsPolicy